refactor(check): extract bindings.ts from checker.ts
Extract variable binding processing functions from checker.ts into a new
bindings.ts module. This is Layer 3 (Constructs) in the refactored
architecture.

Extracted functions:
- processVariableBinding: Main variable binding processor
- emitSimpleBinding: Simple variable binding emission
- extractBindingNodes: Extract nodes from a binding
- emitUnknownTypeError: Emit error for unknown type
- extractBindingIdentInfo: Extract binding identifier info
- isListLiteralBinding: Check if binding is a list literal
- emitListElementBindings: Emit bindings for list elements
- processListLiteralBinding: Process list literal bindings

The record literal case now returns info to the caller (checker.ts) which
handles starting the record literal context, since startRecordLiteral
remains in checker.ts.
